Extract utils and add docs
